feat(ansible): dotfiles decoupling
Dotfiles have now been decoupled and moved into a separate repository
called `brotherkaif/dotfiles`. This allows for dotfiles to be used in a
wider variety of contexts including (but not limited to), GitHub
Codespaces and Docker dev containers. The `dotfiles` play now simply
clones the `dofiles` repo and runs the symlinking script from within
that repository.